Use indices instead of pointers to refer to parent objects
Instead of returning heap-allocated persistent handles, return indices
that map to local handles in the Reader object. This is both a little
faster and makes v8 handle all allocations again.

To make sure that the reader can continue where it left off after
reading an incomplete reply, the local handles are converted to
persistent handles when there was not enough data for a fully reply.
These persistent handles are again converted to local handles on an
subsequent call to Reader::Get().